Menu System
added: menu system
added: screenshot
fixed: hand flip detection
removed: extraneous models
fixed: pencil record sound
modified: pencil to use VRTK events
fixed: load time for loaded scene

+++ Assets/Scripts/MenuManager.cs
@@ -9,6 +9,7 @@ public class MenuManager : MonoBehaviour
     [Range (0,5)]
     public float menuDelay = 1.0f;
     public GameObject menuObject;
+
     #endregion
 
     private bool loadedTextures = false;
@@ -16,23 +17,22 @@ public class MenuManager : MonoBehaviour
     private float t = 0;
     private bool menuDisplaying = false;
     private float angleThreshold = 25;
+    private bool resetting = false;
     GameObject[] menuItems = new GameObject[3];
 
     // Use this for initialization
     void Start()
     {
-        GetComponent<VRTK_ControllerEvents>().AliasMenuOn += new ControllerInteractionEventHandler(HandleMenuPressed);
-        GetComponent<VRTK_ControllerEvents>().AliasMenuOff += new ControllerInteractionEventHandler(HandleMenuReleased);
-
         colorableObjects = FindObjectsOfType<Colorable>();
     }
 
     private void Update()
     {
-        if ((Mathf.Abs(VRTK_DeviceFinder.GetControllerLeftHand(true).transform.localRotation.eulerAngles.x) > 360 - angleThreshold ||
-             Mathf.Abs(VRTK_DeviceFinder.GetControllerLeftHand(true).transform.localRotation.eulerAngles.x) < angleThreshold) &&
-            (Mathf.Abs(VRTK_DeviceFinder.GetControllerLeftHand(true).transform.localRotation.eulerAngles.z) > 180 - angleThreshold &&
-             Mathf.Abs(VRTK_DeviceFinder.GetControllerLeftHand(true).transform.localRotation.eulerAngles.z) < 180 + angleThreshold))
+        Vector3 rot = VRTK_DeviceFinder.GetModelAliasController(gameObject).transform.rotation.eulerAngles;
+        if ((Mathf.Abs(rot.x) > 360 - angleThreshold ||
+             Mathf.Abs(rot.x) < angleThreshold) &&
+            (Mathf.Abs(rot.z) > 180 - angleThreshold &&
+             Mathf.Abs(rot.z) < 180 + angleThreshold))
         {
             t += Time.deltaTime;
         }
@@ -53,65 +53,42 @@ public class MenuManager : MonoBehaviour
         }
     } 
 
-    private void HandleMenuPressed(object sender, ControllerInteractionEventArgs e)
-    {
-        Debug.Log("pressed");
-    }
-
-    private void HandleMenuReleased(object sender, ControllerInteractionEventArgs e)
-    {
-        Debug.Log("released");
-        StartCoroutine(ResetScene());
-    }
-
     public IEnumerator ResetScene()
     {
-        string fullPath = System.IO.Directory.GetCurrentDirectory() + "\\SavedCanvas\\";
-        if (System.IO.Directory.Exists(fullPath))
+        if (!resetting)
         {
-            System.IO.Directory.Delete(fullPath, true);
-        }
+            Debug.Log("resetting");
+            resetting = true;
+            string fullPath = System.IO.Directory.GetCurrentDirectory() + "\\SavedCanvas\\";
+            if (System.IO.Directory.Exists(fullPath))
+            {
+                System.IO.Directory.Delete(fullPath, true);
+            }
 
-        foreach (Colorable co in colorableObjects)
-        {
-            Debug.Log(co.name);
-            VRTK.VRTK_DeviceFinder.GetControllerRightHand(true).GetComponent<ColorObject>().CreateCanvasBase(co);
-            yield return null;
-        }
-        foreach (Colorable co in colorableObjects)
-        {
-            co.canvascam.GetComponent<Camera>().Render();
-            yield return null;
-        }
-    }
-
-    private IEnumerator LoadTexturesFromFile()
-    {
-        string fullPath = System.IO.Directory.GetCurrentDirectory() + "\\SavedCanvas\\";
-        if (System.IO.Directory.Exists(fullPath))
-        {
-            System.DateTime date = System.DateTime.Now;
-            uint i = 0;
-            byte[] bytes;
-            string filename = "";
             foreach (Colorable co in colorableObjects)
             {
-                filename = "CanvasTexture-" + co.name + "-" + i + ".png";
-                i++;
-                if (System.IO.File.Exists(fullPath + filename))
-                {
-                    bytes = System.IO.File.ReadAllBytes(fullPath + filename);
-                    Texture2D tex = new Texture2D(1024, 1024);
-                    tex.LoadImage(bytes);
-                    co.canvasbase.GetComponent<Renderer>().material.mainTexture = tex;
-                    co.canvascam.GetComponent<Camera>().Render();
-                }
+                co.ResetBase();
+                yield return null;
             }
+        }
+        resetting = false;
+    }
+
+    public IEnumerator Screenshot()
+    {
+        string fullPath = System.IO.Directory.GetCurrentDirectory() + "\\Screenshots\\";
+
+        if (!System.IO.Directory.Exists(fullPath))
+            System.IO.Directory.CreateDirectory(fullPath);
+
+        string fileName = "TIAF-Screenshot";
+        int i = 1;
+        while (System.IO.File.Exists(fullPath + fileName + "-" + i + ".png"))
+        {
+            i++;
             yield return null;
         }
-
-        loadedTextures = true;
-
+        Application.CaptureScreenshot(fullPath + fileName + "-" + i + ".png");
     }
 
     private IEnumerator DisplayMenu()
@@ -134,11 +111,11 @@ public class MenuManager : MonoBehaviour
 
         while (menuDisplaying)
         {
-            menuItems[0].transform.localPosition = new Vector3(Mathf.Lerp(0,     0f, lt), Mathf.Lerp(-0.05f, -0.27f, lt), Mathf.Lerp(0,  0.04f, lt));
-            menuItems[0].transform.localScale = Vector3.one * Mathf.Lerp(0.3f, 1.0f, lt);
-            menuItems[1].transform.localPosition = new Vector3(Mathf.Lerp(0,  0.09f, lt), Mathf.Lerp(-0.05f, -0.21f, lt), Mathf.Lerp(0, -0.02f, lt));
-            menuItems[1].transform.localScale = Vector3.one * Mathf.Lerp(0.3f, 1.1f, lt);
-            menuItems[2].transform.localPosition = new Vector3(Mathf.Lerp(0, -0.04f, lt), Mathf.Lerp(-0.05f, -0.14f, lt), Mathf.Lerp(0,  0.11f, lt));
+            menuItems[0].transform.localPosition = new Vector3(Mathf.Lerp(0, 0.09f, lt), Mathf.Lerp(-0.05f, -0.21f, lt), Mathf.Lerp(0, -0.02f, lt));
+            menuItems[0].transform.localScale = Vector3.one * Mathf.Lerp(0.3f, 1.1f, lt);
+            menuItems[1].transform.localPosition = new Vector3(Mathf.Lerp(0,     0f, lt), Mathf.Lerp(-0.05f, -0.27f, lt), Mathf.Lerp(0,  0.04f, lt));
+            menuItems[1].transform.localScale = Vector3.one * Mathf.Lerp(0.3f, 1.0f, lt);
+            menuItems[2].transform.localPosition = new Vector3(Mathf.Lerp(0, -0.04f, lt), Mathf.Lerp(-0.05f, -0.14f, lt), Mathf.Lerp(0, 0.11f, lt));
             menuItems[2].transform.localScale = Vector3.one * Mathf.Lerp(0.3f, 1.2f, lt);
 
             deltat += Time.deltaTime;
@@ -167,12 +144,12 @@ public class MenuManager : MonoBehaviour
 
         while (lt < 1)
         {
-            menuItems[0].transform.localPosition = new Vector3(Mathf.Lerp(menuPos[0].x, 0, lt), Mathf.Lerp(menuPos[0].y, -0.05f, lt), Mathf.Lerp(menuPos[0].z, 0, lt));
-            menuItems[0].transform.localScale = Vector3.one * Mathf.Lerp(1, 0.3f, lt);
-            menuItems[1].transform.localPosition = new Vector3(Mathf.Lerp(menuPos[1].x, 0, lt), Mathf.Lerp(menuPos[1].y, -0.05f, lt), Mathf.Lerp(menuPos[1].z, 0, lt));
-            menuItems[1].transform.localScale = Vector3.one * Mathf.Lerp(1.1f, 0.3f, lt);
+            menuItems[0].transform.localPosition = new Vector3(Mathf.Lerp(menuPos[1].x, 0, lt), Mathf.Lerp(menuPos[1].y, -0.05f, lt), Mathf.Lerp(menuPos[1].z, 0, lt));
+            menuItems[0].transform.localScale = Vector3.one * Mathf.Lerp(1.1f, 0.3f, lt);
+            menuItems[1].transform.localPosition = new Vector3(Mathf.Lerp(menuPos[0].x, 0, lt), Mathf.Lerp(menuPos[0].y, -0.05f, lt), Mathf.Lerp(menuPos[0].z, 0, lt));
+            menuItems[1].transform.localScale = Vector3.one * Mathf.Lerp(1, 0.3f, lt);
             menuItems[2].transform.localPosition = new Vector3(Mathf.Lerp(menuPos[2].x, 0, lt), Mathf.Lerp(menuPos[2].y, -0.05f, lt), Mathf.Lerp(menuPos[2].z, 0, lt));
-            menuItems[2].transform.localScale = Vector3.one * Mathf.Lerp(1, 0.3f, lt);
+            menuItems[2].transform.localScale = Vector3.one * Mathf.Lerp(1.2f, 0.3f, lt);
 
             deltat += Time.deltaTime;
             lt = deltat / transtime;
